requirements.txt
google-generativeai
streamlit
instaloader
